Create directories for each chapter.
Many exercises use the helper.scm file, which belongs in the root of
the project. These exercise load the file, assuming the path structure.
That is it traverses to the parent to load the file. Exercise needed
to load the file one step further up in the hierarchy.

For e.g, instead of (load "../helpers.scm"), it becomes (load "../../helpers.scm")